pyproject.toml
[project]
name = "ulid-dates"
version = "0.1.0"
description = "Determine ULID prefixes for date ranges"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "py-ulid>=1.0.3",
]

[tool.pytest.ini_options]
pythonpath = "src"
